lxd skeleton finished
