--- src/Mysiar/LxdBundle/Model/Lxd.php
@@ -1,17 +1,0 @@
-<?php
-declare(strict_types=1);
-
-/**
- * (c) Piotr Synowiec <psynowiec@gmail.com>
- *
- * 21/10/2017 09:31
- */
-
-namespace Mysiar\LxdBundle\Model;
-
-
-class Lxd
-{
-//    private
-
-}
