+++ src/Mysiar/LxdBundle/README.md
@@ -2,4 +2,13 @@ LxdBundle
 =========
 
 Managing LXD server through API-Platform
-![lxd.info.response](lxd.info.response.png)
+
+This is just skeleton (basically "scouting" the subject) for further development
+
+Example of LXD server response @ https://server:port/1.0
+
+![lxd.info.response](Resources/img/lxd.info.response.png)
+
+Example of LXD container list
+
+![lxd.containers.response](Resources/img/lxd.containers.response.png)
