+++ src/Mysiar/LxdBundle/Resources/config/api_resources/Containers.yml
@@ -1,9 +1,9 @@
-Mysiar\LxdBundle\Model\Info\Info:
+Mysiar\LxdBundle\Model\Containers\Containers:
     properties:
         statusCode:
             identifier: true
     collectionOperations:
         get:
             method: 'GET'
-            route_name: 'Mysiar\LxdBundle\Action\Info'
+            route_name: 'Mysiar\LxdBundle\Action\Containers\GetCollection'
     itemOperations: []
