--- src/Mysiar/LxdBundle/Resources/config/api_resources/Lxd.yml
@@ -1,12 +1,0 @@
-#Mysiar\LxdBundle\Model\Lxd:
-#    collectionOperations:
-#        get:
-#            method: 'GET'
-#            route_name: 'Mysiar\LxdBundle\Action\Info'
-#    itemOperations: []
-#
-#
-#Mysiar\LxdBundle\Model\Info:
-#    properties:
-#        id:
-#          identifier: true
